Change `tests` directory structure so it mirrors `src` structure
