+++ tests/modules/Instabug.spec.js
@@ -7,16 +7,16 @@ import 'react-native';
 import React from 'react';
 import { NativeModules, Platform, processColor, findNodeHandle, Text } from 'react-native';
 import waitForExpect from 'wait-for-expect';
-import './jest/mockInstabug';
-import './jest/mockInstabugUtils';
-import './jest/mockXhrNetworkInterceotor';
-import Instabug from '../src/modules/Instabug';
-import Report from '../src/models/Report';
+import '../jest/mockInstabug';
+import '../jest/mockInstabugUtils';
+import '../jest/mockXhrNetworkInterceotor';
+import Instabug from '../../src/modules/Instabug';
+import Report from '../../src/models/Report';
 
-import IBGConstants from '../src/utils/InstabugConstants';
-import IBGEventEmitter from '../src/utils/IBGEventEmitter';
+import IBGConstants from '../../src/utils/InstabugConstants';
+import IBGEventEmitter from '../../src/utils/IBGEventEmitter';
 import { green } from 'ansi-colors';
-import InstabugUtils from '../src/utils/InstabugUtils';
+import InstabugUtils from '../../src/utils/InstabugUtils';
 
 const { Instabug: NativeInstabug } = NativeModules;
 
@@ -87,7 +87,6 @@ describe('Instabug Module', () => {
   });
 
   it('should not call the native method setIBGLogPrintsToConsole when platform is android', () => {
-
     Platform.OS = 'android';
     Instabug.setIBGLogPrintsToConsole(true);
